requirements.txt
asgiref==3.2.10
Babel==2.8.0
confusable-homoglyphs==3.2.0
countryinfo==0.1.2
Django==3.0.7
django-crispy-forms==1.9.1
django-environ==0.4.5
django-formtools==2.2
django-phonenumber-field==4.0.0
django-recaptcha==2.0.6
django-registration==3.1
phonenumberslite==8.12.6
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
